# gen/http/example/server/encode_decode.go
// Code generated by goa v3.16.2, DO NOT EDIT.
//
// example HTTP server encoders and decoders
//
// Command:
// $ goa gen example/design

package server

import (
	"context"
	"errors"
	example "example/gen/example"
	"io"
	"net/http"

	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeCreateSomeDataResponse returns an encoder for responses returned by
// the example CreateSomeData endpoint.
func EncodeCreateSomeDataResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*example.SomeData)
		enc := encoder(ctx, w)
		body := NewCreateSomeDataResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeCreateSomeDataRequest returns a decoder for requests sent to the
// example CreateSomeData endpoint.
func DecodeCreateSomeDataRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (any, error) {
	return func(r *http.Request) (any, error) {
		var (
			body CreateSomeDataRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return nil, gerr
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateCreateSomeDataRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewCreateSomeDataPayload(&body)

		return payload, nil
	}
}

// EncodeCreateSomeDataError returns an encoder for errors returned by the
// CreateSomeData example endpoint.
func EncodeCreateSomeDataError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "BadRequest":
			var res *goa.ServiceError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewCreateSomeDataBadRequestResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}