# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Example Service","description":"Example service for setting up a new service","version":"0.0.1"},"servers":[{"url":"http://0.0.0.0:80/api/example"}],"paths":{"/api/example":{"post":{"tags":["example"],"summary":"CreateSomeData example","operationId":"example#CreateSomeData","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSomeDataRequestBody"},"example":{"content":"Et repellendus vero.","id":"Quo et."}}}},"responses":{"200":{"description":"Created some data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SomeData"},"example":{"content":"Consequatur occaecati excepturi eius exercitationem repellendus quidem.","created_at":"1983-05-11T10:50:57Z","id":"Voluptatem temporibus."}}}},"400":{"description":"BadRequest: Failed to create some data","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/swagger.json":{"get":{"tags":["example"],"summary":"Download ./gen/http/openapi.json","operationId":"example#/swagger.json","responses":{"200":{"description":"File downloaded"}}}}},"components":{"schemas":{"CreateSomeDataRequestBody":{"type":"object","properties":{"content":{"type":"string","example":"Sit magni quam modi."},"id":{"type":"string","example":"Qui aut dolor iure et."}},"example":{"content":"Consequuntur molestiae et.","id":"Beatae placeat."},"required":["id","content"]},"Error":{"type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"SomeData":{"type":"object","properties":{"content":{"type":"string","example":"Est similique nobis odit."},"created_at":{"type":"string","example":"1970-12-27T09:11:49Z","format":"date-time"},"id":{"type":"string","example":"Qui sint."}},"example":{"content":"Assumenda eaque error.","created_at":"1991-01-09T19:02:05Z","id":"Voluptatem excepturi error."},"required":["id","content","created_at"]}}},"tags":[{"name":"example","description":"The example service is documentation for setting up a new service"}]}